Writable table entries using `@ref` annotation. Table architecture, storage design, and per-entry locks can be configured using `@tableconfig` annotation. See README for further information

+++ src/hardware_dep/shared/includes/ternary_naive.h
@@ -27,10 +27,10 @@ typedef struct {
 typedef struct {
     void**  entries;
     uint8_t keylen;
-    uint8_t size;
+    uint64_t size;
 } ternary_table;
 
-ternary_table* naive_ternary_create (uint8_t keylen, uint8_t max_size);
+ternary_table* naive_ternary_create (uint8_t keylen, uint64_t max_size);
 void           naive_ternary_destroy(ternary_table* t);
 void           naive_ternary_add    (ternary_table* t, uint8_t* key, uint8_t* mask, uint8_t* value);
 uint8_t*       naive_ternary_lookup (ternary_table* t, uint8_t* key);
